remove type field from view nodes
This removes the type field from view nodes and removes normalization
from nodes during compilation. It prevents inconsistencies when using
`data` fields in default components because they won't support the
"class" -> "className" or "for" -> "htmlFor" transformation. Instead of
that, the whole concept of types was removed because it was mainly to
distinguish between components and nodes, but now it isn't needed.
Instead, the type field is gone and the patch checks for name === "text"
instead. Also, there are individual checks for "class" and "for" that
set, update, or remove their corresponding properties.