remove initial data option from data driver
This is a change to make drivers consistent. They should only be an
interface between an application and the browser. In this case, the
browser would have a way of storing things in memory (a variable), and
the data driver provides an interface to changing it. This means there
should not be an initial data option.